CHANGES: - Fixed categories loading from API - Split Categories to 2 separate classes and put them in the 'models' package - Dropdownlist for all job categories shows
