+++ app/src/main/java/com/codepath/synkae/cst438_proj1/RemotiveAPI.java
@@ -1,6 +1,6 @@
 package com.codepath.synkae.cst438_proj1;
 
-import java.util.List;
+import com.codepath.synkae.cst438_proj1.models.Categories;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
